[refactor] Circular in Filter, deps
